microservices/authentication-service/pyproject.toml
[tool.poetry]
name = "authentication-service"
version = "0.1.0"
description = ""
authors = ["Ivan Gruzdev <gruzdev.ivan.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
pydantic-settings = "^2.4.0"
grpcio-tools = "^1.65.4"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
grpcio = "^1.65.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
